+++ src/problems/problem_038.rs
@@ -13,16 +13,16 @@ fn problem_038() {
     println!("{}", result);
 }
 
-fn is_pandigital(digits: &Vec<i64>) -> bool {
+fn is_pandigital(digits: &Vec<i8>) -> bool {
     if digits.len() != 9 {
         return false;
     }
 
-    let uniq_digits = digits.into_iter().collect::<HashSet<&i64>>();
+    let uniq_digits = digits.into_iter().collect::<HashSet<&i8>>();
     uniq_digits.len() == 9 && !uniq_digits.contains(&0)
 }
 
-fn concatenated_product(num: i64) -> Vec<i64> {
+fn concatenated_product(num: i64) -> Vec<i8> {
     let mut digits = vec![];
     for i in 1..=9 {
         digits.append(&mut (num * i).digits());
